+++ Content.Shared/Nyanotrasen/Abilities/Psionics/GuaranteedPsionicComponent.cs
@@ -1,4 +1,4 @@
-namespace Content.Server.Psionics
+namespace Content.Shared.Abilities.Psionics
 {
     [RegisterComponent]
     public sealed class GuaranteedPsionicComponent : Component
